# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Home",
        "version": "1.0.0"
    },
    "paths": {
...
        "/guilds/balance": {
            "post": {
                "tags": [
                    "Guilds"
                ],
                "summary": "Balancear guildas",
                "operationId": "2cb4c239a41b7d1ed7a357ca7b5a116b",
                "responses": {
                    "200": {
                        "description": "Guildas balanceadas com sucesso"
                    },
                    "500": {
                        "description": "Erro ao balancear guildas"
                    }
                }
            }
        },
        "/guilds": {
            "get": {
                "operationId": "6da9fe90b35488a3597f1a3d106655b3",
                "responses": {
                    "200": {
                        "description": "Lista de guildas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guild"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Guilds"
                ],
                "summary": "Criar uma nova guilda",
                "operationId": "cc5542c43e53c3b178a97ade476c0ade",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Guild"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guilda criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guild"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar guilda"
                    }
                }
            }
        },
        "/guilds/{id}": {
            "get": {
                "tags": [
                    "Guilds"
                ],
                "summary": "Exibir uma guilda específica",
                "operationId": "4f8c55cf7e2ff2c059f6ceba4e4ec0b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da guilda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guilda encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guild"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guilda não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guilds"
                ],
                "summary": "Atualizar uma guilda específica",
                "operationId": "f3622a7489f2f5af59b48d21762431b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da guilda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Guild"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guilda atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guild"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar guilda"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guilds"
                ],
                "summary": "Deletar uma guilda específica",
                "operationId": "4aac1e4eb94350fea3ffd6205f1a04a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da guilda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guilda deletada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao deletar guilda"
                    }
                }
            }
        },
...
        "/login": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Exibe a página de login.",
                "operationId": "c6c57832bcce09c6e5a4cd562f1fedc5",
                "responses": {
                    "200": {
                        "description": "Retorna a página de login."
                    }
                }
            },
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Processa o login do usuário.",
                "operationId": "aec6d9fd8be3c8b24cda0c17a60a8bb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erro de autenticação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "Realiza o logout do usuário.",
                "operationId": "214cdf6ce10b65006d77963647f563fb",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao realizar logout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno do servidor ao realizar logout."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Guild": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Guild Name"
                    },
                    "level": {
                        "type": "integer",
                        "example": 10
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Player1",
                            "Player2"
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
...
}